Take value of override_url into account when assigning URLs to pages
It is possible to move an entire page tree to another base URL now. If
you need a home page where every other page is a child, assign the home
page a override_url of '/'.

While at it, simplify page_for_path to use _cached_url instead of doing
all the work again.