Implement a post_syncdb listener which checks database schema consistency
Since south cannot do non-linear and optional migrations (which would be
necessary to activate and deactivate page extensions) we try at least to
be a little more helpful than simply crashing upon non-existing columns.

The user is informed upon syncdb that several columns might be missing.

Please note that this has not been tested yet with ManyToManyFields --
they obviously do not show up when inspecting the primary table, and the
code does not currently handle this.