Implement an alternative approach for integrating 3rd party apps
This approach has the advantage that the point where the application will be
available can be freely defined in the page editor. On the other hand, the
integration is less flexible and modifying the third-party application
cannot be avoided. Plus, it monkey-patches django.core.urlresolvers.reverse,
which can be seen as ugly (and I'd say rightly so).

Still, it might be more than a nice hack...