Admin Performance Boost
Previously, the tree editor would perform O(n) - which was slow in production and unusable with DEBUG=True with a test database over 50+ pages (I have 2500). This uses .select_related() and .only("id", "parent") to reduce that to one query. A second, better rewrite would use .values_list() to avoid paying the memory tax of loading thousands of complex objects.