Admin Performance Boost #2
Changed the TreeEditorQuerySet iterator to use Django 1.1 aggregates to pull in the entire tree with one query rather than one per leaf. In my test data this reduced teh total number of queries for the admin page from 172 to 9, and the load time from ~8 seconds to 4 (which is still high but closer to tolerable).