Frontend editing fix: CSS position
The position:absolute, bottom:-25px element was unselectable because the
mouseout would fire if you moved the mouse over the edit element. I
moved it to the upper right corner of the content box to fix this and
also avoid the edit link being "hidden" at the bottom of a large content
page.