Experimental Admin support for Django reversion
Admin change view now supports a "reversion=<id>" GET parameter which causes that revision to be loaded as if it were form POST data. This populates the admin interface but allows the user to make changes before saving.

Added a messages list for the view so reversion can produce a nice admin-style list at the top with a helpful "Click save to overwrite current" display.