Admin Category FilterSpec improvements: only displays used categories
Previous all categories on the system would be displayed; now the list is limited to those which are actually in use by the model in question. This produces a much smaller list if either not all categories are in use or if the same category list is shared across multiple models.