Implement different language selection strategies.
Add setting FEINCMS_TRANSLATION_POLICY for selecting the selection strategies.
Make the translation extension load different request processors depending on the strategy.
Implement an "EXPLICIT" strategy that will switch a user's language only if there is no preference in place or the language is explicitely switched. With this strategy, one can navigate on multi-language pages without the site language being switched to and fro all the time.