Adapt FeinCMS to upcoming django-mptt 0.4 release
Multiple inheritance of Django models still seems to be a little
shaky. Therefore we better work with a linearized inheritance
tree. django-mptt has been changed to take advantage of abstract
models (which is the right thing to do) but this forces us to
jump through a few hoops.

The whitespace-ignoring diff (git diff -b) of this commit looks
much less scary by the way.