Rework ct_tracker Page extension to fix issue #113.
Previously, we just computed the set of content types in a page in the pre_save handler. However, at the time the pre_save signal is sent, the corresponding foreign key tables have not yet been updated. Therefore the ct_tracker was always "one save behind" wrt to the real content of the page.

Using post_save does not change that, the fieldset tables are obviously updated even later than that by django.admin.

Therefore, we delay updating the _ct_inventory field: just null out the field on save and rebuild it whenever a page with an empty inventory is accessed.