Fine tune interaction between application content and remote caching.
Since application content usually do not want to be cached by clients or
proxies, make sure that a page with an app content will not set the
Last-Modified header to the page's last modified date. Also, add a
Cache-Control header to pages with an app content that instructs not
to cache this page.