Allow configuring a cache key prefix for application contents
We use our own Page module, which mostly overlaps with the standard
model but implements multi-language capabilities (especially with
regards to _cached_url). This generally works very robustly but breaks
when using the application content module.

What happens is that app_reverse() goes ahead and caches the results of
get_absolute_url() for a page without regard for the currently set
language. This is a problem for us, as the result of get_absolute_url()
changes whenever the current language changes.

This commit introduces a setting to configure a function that is called
whenever a cache key is generated for the reverse cache. This function
must return a string to be prefixed to the cache key.