Decouple models and views some more
Move the processor and content type handling from Page and Handler into
reusable classes.

This commit is essentially a big move of code and should not have caused
any change of behavior. Unfortunately it did: It moves the execution of
request processors where it belongs (in the view), which means that they
aren't run in Page.setup_request anymore, which in turn causes the
require_path_active_request_processor to NOT run inside the page manager
methods, which means that now inactive pages may be returned from the
following methods:

- Page.objects.page_for_path
- Page.objects.best_match_for_path
- Page.for_request

The next commit tries rectifying that.

Refs #241.