Exclude pages with inactive parents in manager methods
The old handling where the activity check was run in code which
belongs to the request-response cycle (but still was inside the
models file!) was arguably broken.

This commit adds a new argument to all page manager methods,
require_path_active, which replaces the functionality of
require_path_active_request_processor.