added support for changing quality in thumbnails
The fincms_thumbnail tags are used for much more than generating small
thumbs. I have added the option of setting the quality to save
bandwidth on image heavy pages. It can be set in the settings file or
on a per-use basis which will override the value in settings. This will
cause all images to regenerate as quality is now part of the filename.