get_navigation_url could never have worked right: When having an internal link "page.page:13", it would return the url of the current page.
Even if the comparison had been right, then the returned url would have been "page.page:13", which is totally bogus for something to put into an <a> tag. Now tries to resolve the redirect into a real url.

Also factor out a get_redirect_to_target(), so that one can get the real internal target of a page in templates too.