refactor SQL queries to use django queryset API, fixes #621
I did need to effectively comment out parts of  because it was expecting a certain number of SQL queries (4) and the changes now bump this up to 14. The trade-off is there are a few more queries, but the results are cached, and the code is a little cleaner. Thoughts on how to resolve the tests are welcome.