Close image after reading dimensions. The mediafile admin page triggers a ResourceWarning;
The warning you get to see after viewing a mediafile webpage is::

	ResourceWarning: unclosed file <_io.BufferedReader name='/***/helloworld.png'>
	ResourceWarning: Enable tracemalloc to get the object allocation traceback

Django's get_image_dimensions() leaves the file open unless you call
close=True; Alternatively the admin code could do add a finally clause to the
try/except and do a obj.file.close() there. But this change seems to be in the
spirit of get_image_dimensions().